+++ dokdo/api/beta_parallel_plot.py
@@ -46,10 +46,10 @@ def beta_parallel_plot(
 
     See Also
     --------
-    ordinate
-    beta_2d_plot
-    beta_3d_plot
-    beta_scree_plot
+    dokdo.api.ordinate
+    dokdo.api.beta_2d_plot
+    dokdo.api.beta_3d_plot
+    dokdo.api.beta_scree_plot
 
     Examples
     --------