+++ dokdo/api/ordinate.py
@@ -50,10 +50,10 @@ def ordinate(
 
     See Also
     --------
-    beta_2d_plot
-    beta_3d_plot
-    beta_scree_plot
-    beta_parallel_plot
+    dokdo.api.beta_2d_plot
+    dokdo.api.beta_3d_plot
+    dokdo.api.beta_scree_plot
+    dokdo.api.beta_parallel_plot
 
     Notes
     -----