+++ dokdo/api/regplot.py
@@ -40,7 +40,7 @@ def regplot(
 
     See Also
     --------
-    taxa_abundance_box_plot
+    dokdo.api.taxa_abundance_box_plot
 
     Examples
     --------